Add useGetExercisesByCollectionId
